* Free lexemes clobbered as happaxes

+++ spacy/_hashing.pxd
@@ -6,6 +6,6 @@ cdef class FixedTable:
     cdef uint64_t* keys
     cdef size_t* values
 
-    cdef int insert(self, uint64_t key, size_t value) nogil
+    cdef size_t insert(self, uint64_t key, size_t value) nogil
     cdef size_t get(self, uint64_t key) nogil
     cdef int erase(self, uint64_t key) nogil
